+++ package/Editor/Sources/UI/Common/Alert.cs
@@ -68,7 +68,10 @@ namespace UnityEditor.PackageManager.UI
                 style.positionRight = originalPositionRight;
         }
 
-        private Label AlertMessage { get { return root.Q<Label>("alertMessage"); } }
-        private Button CloseButton { get { return root.Q<Button>("close"); } }
+        private Label _alertMessage;
+        private Label AlertMessage { get { return _alertMessage ?? (_alertMessage = root.Q<Label>("alertMessage")); } }
+
+        private Button _closeButton;
+        private Button CloseButton { get { return _closeButton ?? (_closeButton = root.Q<Button>("close")); } }
     }
 }