+++ package/Editor/Sources/UI/PackageManagerPrefs.cs
@@ -1,12 +1,15 @@
-using UnityEngine;
-using UnityEditorInternal;
+using System;
+using System.IO;
+using System.Security.Cryptography;
+using System.Text;
 
 namespace UnityEditor.PackageManager.UI
 {
-    internal class PackageManagerPrefs
+    internal static class PackageManagerPrefs
     {
         private const string showPreviewPackagesPrefs = "PackageManager.ShowPreviewPackages";
         private const string showPreviewPackagesWarningPrefs = "PackageManager.ShowPreviewPackagesWarning";
+        private const string lastUsedFilterPrefix = "PackageManager.Filter_";
 
         public static bool ShowPreviewPackages
         {
@@ -18,6 +21,23 @@ namespace UnityEditor.PackageManager.UI
         {
             get { return EditorPrefs.GetBool(showPreviewPackagesWarningPrefs, true); }
             set { EditorPrefs.SetBool(showPreviewPackagesWarningPrefs, value); }
+        }
+
+        private static string GetHascodeHexString(string str)
+        {
+            var bytes = new SHA1Managed().ComputeHash(Encoding.UTF8.GetBytes(Directory.GetCurrentDirectory()));
+            return BitConverter.ToString(bytes);            
+        }
+
+        public static PackageFilter GetLastUsedPackageFilter(string str)
+        {
+            return (PackageFilter)Enum.Parse(typeof(PackageFilter),
+                EditorPrefs.GetString(lastUsedFilterPrefix + GetHascodeHexString(str), PackageFilter.All.ToString()));
+        }
+
+        public static void SetLastUsedPackageFilter(string str, PackageFilter filter)
+        {
+            EditorPrefs.SetString(lastUsedFilterPrefix + GetHascodeHexString(str), filter.ToString());
         }
     }
 }